Rename "convert.h" to "event_converter.h"
The filename gave no hint about what was converted.
