+++ app/meson.build
@@ -3,10 +3,10 @@ src = [
     'src/command.c',
     'src/control_msg.c',
     'src/controller.c',
-    'src/convert.c',
     'src/decoder.c',
     'src/device.c',
     'src/device_msg.c',
+    'src/event_converter.c',
     'src/file_handler.c',
     'src/fps_counter.c',
     'src/input_manager.c',
