+++ app/src/input_manager.c
@@ -1,7 +1,7 @@
 #include "input_manager.h"
 
 #include <SDL2/SDL_assert.h>
-#include "convert.h"
+#include "event_converter.h"
 #include "lock_util.h"
 #include "log.h"
